refactor: Styling refactor
Moved all css styling to singular external css file
