--- src/components/Checklist/ChecklistMeta.css
@@ -1,3 +1,0 @@
-.meta {
-    color: #656565;
-}
