+++ src/components/Checklist/ChecklistMeta.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import { FormattedTime, FormattedDate, FormattedMessage } from 'react-intl';
-import css from './ChecklistMeta.css';
+import css from './Checklist.css';
 
 const propTypes = {
   dateCreated: PropTypes.string,
