--- src/components/Checklist/ChecklistNotesForm.css
@@ -1,7 +1,0 @@
-.container {
-    padding: 1rem;
-}
-
-.button {
-    padding-left: 1rem;
-}
